Moved the multi enum class
